[RND-364] Refactor test structure
