+++ Meadowlark-js/tests/e2e/EducationOrganizations.test.ts
@@ -3,7 +3,8 @@
 // The Ed-Fi Alliance licenses this file to you under the Apache License, Version 2.0.
 // See the LICENSE and NOTICES files in the project root for more information.
 
-import { baseURLRequest, Clients, getAccessToken } from './SharedFunctions';
+import { getAccessToken, Clients } from './functions/Credentials';
+import { baseURLRequest } from './Setup';
 
 describe('Education Organizations', () => {
   it("shouldn't allow to post to abstract entity", async () => {
