+++ Meadowlark-js/tests/e2e/Sessions.test.ts
@@ -3,16 +3,10 @@
 // The Ed-Fi Alliance licenses this file to you under the Apache License, Version 2.0.
 // See the LICENSE and NOTICES files in the project root for more information.
 
-import {
-  baseURLRequest,
-  Clients,
-  createResource,
-  createSchool,
-  deleteByLocation,
-  getAccessToken,
-  getDescriptorByLocation,
-  rootURLRequest,
-} from './SharedFunctions';
+import { getAccessToken, Clients } from './functions/Credentials';
+import { createResource, deleteResourceByLocation } from './functions/Resources';
+import { getDescriptorByLocation, createSchool } from './functions/Shared';
+import { baseURLRequest, rootURLRequest } from './Setup';
 
 describe('Sessions', () => {
   describe('with strict validation', () => {
@@ -97,9 +91,9 @@ describe('Sessions', () => {
     });
 
     afterEach(async () => {
-      await deleteByLocation(sessionLocation);
-      await deleteByLocation(termDescriptorLocation);
-      await deleteByLocation(schoolLocation);
+      await deleteResourceByLocation(sessionLocation);
+      await deleteResourceByLocation(termDescriptorLocation);
+      await deleteResourceByLocation(schoolLocation);
     });
   });
 });
