+++ Meadowlark-js/tests/e2e/StudentInterventionAssociations.test.ts
@@ -3,14 +3,9 @@
 // The Ed-Fi Alliance licenses this file to you under the Apache License, Version 2.0.
 // See the LICENSE and NOTICES files in the project root for more information.
 
-import {
-  baseURLRequest,
-  getAccessToken,
-  Clients,
-  deleteByLocation,
-  rootURLRequest,
-  createResource,
-} from './SharedFunctions';
+import { getAccessToken, Clients } from './functions/Credentials';
+import { createResource, deleteResourceByLocation } from './functions/Resources';
+import { baseURLRequest, rootURLRequest } from './Setup';
 
 describe('Student Intervention Association', () => {
   describe('with strict validation', () => {
@@ -69,7 +64,7 @@ describe('Student Intervention Association', () => {
     });
 
     afterAll(async () => {
-      await deleteByLocation(location);
+      await deleteResourceByLocation(location);
     });
   });
 });
