+++ Meadowlark-js/tests/e2e/educationContents.test.ts
@@ -3,17 +3,10 @@
 // The Ed-Fi Alliance licenses this file to you under the Apache License, Version 2.0.
 // See the LICENSE and NOTICES files in the project root for more information.
 
-import {
-  baseURLRequest,
-  rootURLRequest,
-  generateRandomId,
-  getAccessToken,
-  deleteByLocation,
-  createContentClassDescriptor,
-  getDescriptorByLocation,
-  Clients,
-  createResource,
-} from './SharedFunctions';
+import { getAccessToken, Clients } from './functions/Credentials';
+import { createResource, deleteResourceByLocation } from './functions/Resources';
+import { createContentClassDescriptor, getDescriptorByLocation, generateRandomId } from './functions/Shared';
+import { baseURLRequest, rootURLRequest } from './Setup';
 
 describe('Education contents', () => {
   let educationContentLocation: string;
@@ -94,10 +87,10 @@ describe('Education contents', () => {
   });
 
   afterEach(async () => {
-    await deleteByLocation(educationContentLocation);
+    await deleteResourceByLocation(educationContentLocation);
   });
 
   afterAll(async () => {
-    await deleteByLocation(contentClassDescriptorLocation);
+    await deleteResourceByLocation(contentClassDescriptorLocation);
   });
 });
